Removed Radio Buttons, FIxed bugs
Removed the Radio Buttons (too complex to maintain and teach)
Fixed bugs with lists with numbers not working. Fixed bug with text boxes with multiple lines creating syntax error in project files.  Think I fixed the bug with the click file running before the objects it used were defined (not tested).  Tried to make it report syntax errors in click file but it still won't this needs more work.