Mouse wheel update, final
Made the mouse wheel not only change the font but also change the height of the clicks and graphics windows, proportionately.  This way if one needs more space for graphics one can use a smaller font, but if one wants to see demo one can use a larger font.  The windows will always show 15 lines of text now.